ci: refactor into a package
# Conflicts:
#	uv.lock
