+++ .github/workflows/lint.yaml
@@ -47,4 +47,4 @@ jobs:
         run: uv run ruff check
 
       - name: Lint changelog
-        run: uv run ci/lint_changelog.py
+        run: uv run lint-changelog
