+++ ci/src/freecam_ci/project_files.py
@@ -1,0 +1,6 @@
+from pathlib import Path
+
+ROOT = Path(__file__).resolve().parents[3]
+CHANGELOG_FILE = ROOT / "CHANGELOG.md"
+METADATA_FILE = ROOT / "metadata.toml"
+STONECUTTER_FILE = ROOT / "stonecutter.json5"
