+++ ci/src/freecam_ci/stonecutter_model.py
