+++ ci/test/fixtures/non_str_version.toml
